+++ src/input.ts
@@ -4,6 +4,8 @@ import fs from 'fs';
 export interface ChainRef {
     /** PDB entry identifier or equivalent */
     entryId: string,
+    /** ID of the assembly within entry to be processed (or `undefined` to process deposited model) */
+    assemblyId: string | undefined,
     /** auth_asym_id of selected chain (or `undefined` to process all chains) */
     chainId: string | undefined,
 }
@@ -13,8 +15,18 @@ export function loadInputDataset(file: string): ChainRef[] {
     const lines = text.split('\n').map(l => l.trim()).filter(l => l.length > 0 && !l.startsWith('#'));
     const result = [];
     for (const line of lines) {
-        const [entryId, chainId] = line.split(',');
-        result.push({ entryId, chainId });
+        const [struct, chainId] = line.split(/[-,]/);
+        const [entryId, assemblyId] = struct.split('_');
+        result.push({ entryId, assemblyId, chainId });
     }
     return result;
+}
+
+/** Return filename for outputs (without file extension) */
+export function createFilename(chainRef: ChainRef): string {
+    const { entryId, assemblyId, chainId } = chainRef;
+    let out = entryId;
+    if (assemblyId) out += `_${assemblyId}`;
+    if (chainId) out += `-${chainId}`;
+    return out;
 }